Moved everything from the header file other than the load and save functions into the .cpp file. This made it possible to remove lots of forward declarations and also made the header file much easier to understand (more in line with the spirit of deadfrog lib).
Replaced the table based CRC with a tableless one. In theory it is slower but I couldn't detect a difference in my benchmarks - any difference must be less than 1%. The result is smaller source and 1K smaller binary.

Removed some more color conversion stuff because we always want 32-bit bitmaps.